compressed-ddp/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
matplotlib>=3.7.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
